ml-model/requirements.txt
flask
pandas
numpy
scikit-learn
nltk
spacy
pdfminer.six
docx2txt
joblib
